Update version number to 0.0.6 in package.json
Add command to delete traQ message

This commit updates the version number in the package.json file to 0.0.6. It also adds a new command to delete a message in the traQ channel. The command prompts the user to enter the message ID and sends a DELETE request to the traQ API to delete the message. If successful, a success message is displayed, otherwise an error message is shown.